Changed Tkinter's layout manager and changed the control bar
Change the layout manager from pick to place, allowing greater freedom
in the placement of widgets.
Remove the frames and, as a consequence, add the background of the
controls bar (controls_bar.png).
Add the time bar to the control bar (time_bar.png), as well as its mouse
enter and mouse leave events.