Added partial UI elements responsible for submitting comments
Add a canvas.
Add a frame inside the canvas.
Add a vertical scrollbar attached to the canvas.
Explanation: a scrollbar can only be related to one widget, so a canvas
was created and, in order to indirectly relate several widgets with the
scrollbar, a frame was added to the canvas. Later, the widgets for each
comment will be inserted into that frame.