Added code architecture required to handle comments
Change function that handles mouse button 1 click event in ui.py, so
that it receives a function as parameter, which will be executed.
Add function in ui.py responsible for storing the path of the .txt file
selected by the user and call main functions that are responsible for
managing comments from the .txt file.
Add main global variables in main.py.
Add main functions for handling comments in main.py.
Created main_constants.py file and added constants that represent the
possible states of a comment.