Developed functionality that displays the total time in the ui
Add the totalTime attribute to the tuple that contains the time data.
Add function that updates totalTime every time a txt file has its
comments loaded.
Add parameters to ui.py event functions that indicate whether to execute
of them will occur based on the return value of any function and the
required value.
for the execution to take place. This is necessary so that the function
related to the Button-1 event of the play/pause button is only executed
if the current time is not equal to the total time.
Developed function in main.py that checks whether the current time has
reached the total time.
Put "velocity" value in the main tuple that contains the weather data