Event function code refactored and current time input development started by etrCurrentTime
Replace the repeated codes of some of the functions that are called by
events with a single function that contains that code and is called by
the others.
Add functions and constants in ui files to allow changing the current
time by the Entry widget which shows the current time.