Added comments with a weaker color when they were just loaded in the GUI and created a function to change the state of the comments
Add color-coded constant that makes comments look like they're disabled
when they're loaded into the GUI, just like BigBlueButton does.
Created a function in main.py to change the state of a comment, so that
the ui could do it after loading a comment, something that wasn't being
done before.