Added functionality to handle root closure and changed functions that display messages
Add protocol to root so that when the user clicks the close button, it
triggers a function that will or will not present a message asking the
user to confirm the closing of the form.
Change functions that have msgbox by adding two attributes that let you
decide if they will be executed or not.