Changed functions responsible for closing the software
Add the global variable that indicates whether the user has requested
the program to end.
Allow the program to terminate, even when more than one thread is
running, waiting 1 second before terminating it to give time for the
time thread or the thread that carries the comments to terminate.
Indicate in the status bar that the program is ending.